Squashed commit of the following:
commit 307fac12789e3060b8de627b01c7b150726f067c
Merge: 1e7f1c5 f057a7f
Author: Pablo Jimenez <pablo.jimenez-martinez@capgemini.com>
Date:   Mon May 13 08:52:16 2024 +0200

    Merge pull request #7 from ccsw-csd/feature/members_admins

    Feature/members admins Añadir funcionalidad. Crear y Editar grupos con members y admins

commit f057a7f97e247a1ecd2186e5dd795a74394715d2
Author: mguaitav <miguel-angel.guaita-villasenor.external@capgemini.com>
Date:   Fri May 10 12:27:35 2024 +0200

    fix: implementar correcciones de Pablo

commit 41573dae7f286abd182fbd7de42ebc3e3148dce7
Author: mguaitav <miguel-angel.guaita-villasenor.external@capgemini.com>
Date:   Thu May 9 13:25:38 2024 +0200

    fix: corregir los errores del merge

    cambio de Integer a Long el PersonEntity para que el backend funcione, tambien la clase Topic

commit 3eb17829881904efce0b3dd06541d0cf61128425
Merge: d0262b8 22e89e7
Author: mguaitav <miguel-angel.guaita-villasenor.external@capgemini.com>
Date:   Thu May 9 13:06:01 2024 +0200

    Merge branch 'develop' into feature/members_adminsmerge

commit d0262b8d3a2e2b6f46b3344be7ab0e33eb184d76
Author: mguaitav <miguel-angel.guaita-villasenor.external@capgemini.com>
Date:   Thu May 9 12:13:05 2024 +0200

    fix: cambiar todas las cosas que pablo corrigio

    cambiar Longs por Entities en las Entidades complejas (GroupAdmin y GroupStaff
    borrar metodos que no utilizo
    refactorizar metodos que utilizo

commit c6e9d8946bb2094341f2e12042ee1b58a4e54fce
Author: mguaitav <miguel-angel.guaita-villasenor.external@capgemini.com>
Date:   Fri May 3 12:31:38 2024 +0200

    feat. Añadir funcionalidad Grupos - miembros y admins

    se pueden crear y editar grupos con sus miembros y admins
    El usuario que crea un grupo es puesto automáticamente como admin

commit 9881a25ba79484f13e17f08cca4084ce3e0fdaa5
Author: mguaitav <miguel-angel.guaita-villasenor.external@capgemini.com>
Date:   Mon Apr 22 10:35:47 2024 +0200

    feat: añadir primer commit de clase person para revision de pablo

    para revisar el problema de peticiones a db

+++ src/main/java/com/cca/moodmeter/topic/TopicServiceImpl.java
@@ -161,7 +161,7 @@ public class TopicServiceImpl implements TopicService {
         if (data.getStatus() == 1) {
             String user = UserUtils.getUserDetails().getUsername();
             PersonEntity person = personRepository.findByUsernameAndActiveTrue(user);
-            Integer personId = person.getId();
+            Long personId = person.getId();
 
             boolean hasVoted = topicVotedByRepository.existsByTopicIdAndPersonId(data.getId(), personId);
             if (!hasVoted) {
