+++ src/main/java/com/cca/moodmeter/topic/TopicVotedByRepository.java
@@ -6,6 +6,6 @@ import com.cca.moodmeter.topic.model.TopicVotedByEntity;
 
 public interface TopicVotedByRepository extends CrudRepository<TopicVotedByEntity, Long> {
 
-    boolean existsByTopicIdAndPersonId(Long id, Integer personId);
+    boolean existsByTopicIdAndPersonId(Long id, Long personId);
 
 }
